+++ poetry/utils/env.py
@@ -1,6 +1,7 @@
 import json
 import os
 import platform
+import re
 import subprocess
 import sys
 import sysconfig
@@ -15,6 +16,7 @@ from typing import Tuple
 
 from poetry.config import Config
 from poetry.locations import CACHE_DIR
+from poetry.semver.version import Version
 from poetry.utils._compat import Path
 from poetry.utils._compat import decode
 from poetry.utils._compat import encode
@@ -114,6 +116,7 @@ class Env(object):
         self._base = base or path
 
         self._marker_env = None
+        self._pip_version = None
 
     @property
     def path(self):  # type: () -> Path
@@ -151,6 +154,25 @@ class Env(object):
         Path to current pip executable
         """
         return self._bin("pip")
+
+    @property
+    def pip_version(self):
+        if self._pip_version is None:
+            self._pip_version = self.get_pip_version()
+
+        return self._pip_version
+
+    @property
+    def site_packages(self):  # type: () -> Path
+        if self._is_windows:
+            return self._path / "Lib" / "site-packages"
+
+        return (
+            self._path
+            / "lib"
+            / "python{}.{}".format(*self.version_info[:2])
+            / "site-packages"
+        )
 
     @classmethod
     def get(cls, cwd, reload=False):  # type: (Path, bool) -> Env
@@ -308,6 +330,9 @@ class Env(object):
     def config_var(self, var):  # type: (str) -> Any
         raise NotImplementedError()
 
+    def get_pip_version(self):  # type: () -> Version
+        raise NotImplementedError()
+
     def is_valid_for_marker(self, marker):  # type: (BaseMarker) -> bool
         return marker.validate(self.marker_env)
 
@@ -424,6 +449,11 @@ class SystemEnv(Env):
 
             return
 
+    def get_pip_version(self):  # type: () -> Version
+        from pip import __version__
+
+        return Version.parse(__version__)
+
     def is_venv(self):  # type: () -> bool
         return self._path != self._base
 
@@ -473,6 +503,14 @@ class VirtualEnv(Env):
             value = 0
 
         return value
+
+    def get_pip_version(self):  # type: () -> Version
+        output = self.run("python", "-m", "pip", "--version").strip()
+        m = re.match("pip (.+?)(?: from .+)?$", output)
+        if not m:
+            return Version.parse("0.0")
+
+        return Version.parse(m.group(1))
 
     def is_venv(self):  # type: () -> bool
         return True
@@ -546,6 +584,7 @@ class MockEnv(NullEnv):
         platform="darwin",
         os_name="posix",
         is_venv=False,
+        pip_version="19.1",
         **kwargs
     ):
         super(MockEnv, self).__init__(**kwargs)
@@ -555,6 +594,7 @@ class MockEnv(NullEnv):
         self._platform = platform
         self._os_name = os_name
         self._is_venv = is_venv
+        self._pip_version = Version.parse(pip_version)
 
     @property
     def version_info(self):  # type: () -> Tuple[int]
@@ -571,6 +611,10 @@ class MockEnv(NullEnv):
     @property
     def os(self):  # type: () -> str
         return self._os_name
+
+    @property
+    def pip_version(self):
+        return self._pip_version
 
     def is_venv(self):  # type: () -> bool
         return self._is_venv